+++ .gitignore
@@ -1,11 +1,10 @@
 /local.properties
 /.idea
-.idea
 .DS_Store
 /build
 /captures
 .cxx
-
+/.kotlin
 # Created by https://www.toptal.com/developers/gitignore/api/androidstudio
 # Edit at https://www.toptal.com/developers/gitignore?templates=androidstudio
