+++ app/src/main/java/com/zywczas/weather/di/KoinModule.kt
@@ -1,7 +1,7 @@
 package com.zywczas.weather.di
 
-import com.zywczas.common_utils.commonUtilModule
-import com.zywczas.network_forecast.networkForecastModule
+import com.zywczas.commonutil.commonUtilModule
+import com.zywczas.networkforecast.networkForecastModule
 import com.zywczas.networkcaller.networkCallerModule
 import com.zywczas.weather.screens.searchcity.SearchCityViewModel
 import org.koin.core.module.dsl.viewModelOf
