+++ common_util/src/main/java/com/zywczas/commonutil/Loggers.kt
@@ -1,4 +1,4 @@
-package com.zywczas.common_utils
+package com.zywczas.commonutil
 
 import android.util.Log
