+++ common_util/src/main/java/com/zywczas/commonutil/StringProvider.kt
@@ -1,4 +1,4 @@
-package com.zywczas.common_utils
+package com.zywczas.commonutil
 
 import android.content.Context
 import androidx.annotation.StringRes
