+++ network_caller/src/main/java/com/zywczas/networkcaller/ApiKeys.kt
@@ -1,7 +1,5 @@
 package com.zywczas.networkcaller
 
-import com.zywczas.network_caller.BuildConfig
-
 object ApiKeys {
     const val OPEN_WEATHER_MAP = BuildConfig.OPENWEATHERMAP_API_KEY
 }
