+++ network_forecast/src/main/java/com/zywczas/networkforecast/openweatherapi/Units.kt
@@ -1,4 +1,4 @@
-package com.zywczas.network_forecast.openweatherapi
+package com.zywczas.networkforecast.openweatherapi
 
 enum class Units(val value: String) {
     Metric("metric")
