+++ network_forecast/src/main/java/com/zywczas/networkforecast/response/CurrentResponse.kt
@@ -1,4 +1,4 @@
-package com.zywczas.network_forecast.response
+package com.zywczas.networkforecast.response
 
 import com.google.gson.annotations.SerializedName
